feat : add os file (#3)
* feat : add os file

* feat : add directory os

* feat : improve test

+++ filesystem/file.go
@@ -7,5 +7,5 @@ type FileOperation interface {
 	WriteAll(bytes []byte) error
 	ReadAll() ([]byte, error)
 	ReadAt(bytes []byte, offset int) (int, error)
-	Size() (int, error)
+	Size() (int64, error)
 }
